# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Grocery AI Store API",
        "description": "AI-Assisted Health & Grocery Store API. Browse and purchase health-focused grocery products with AI-powered product guidance, personalized diet suggestions, and smart product search.",
        "contact": {
            "email": "admin@smartgrocery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/customer/ai/ask": {
            "post": {
                "tags": [
                    "AI Assistant"
                ],
                "summary": "Ask AI a question",
                "description": "Send a question to the AI assistant about products, nutrition, or health. Returns personalized responses based on user's health profile.",
                "operationId": "860ca9191a340800da26eb0323e0439b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "What are good protein sources for muscle gain?"
                                    },
                                    "context": {
                                        "type": "string",
                                        "example": "product_recommendation",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AI response generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "response": {
                                                    "type": "string",
                                                    "example": "Based on your health profile and goals, I recommend..."
                                                },
                                                "suggested_products": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "conversation_id": {
                                                    "type": "integer",
                                                    "example": 42
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "AI response generated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Rate limit exceeded"
                    },
                    "500": {
                        "description": "AI service error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/ai/suggestions": {
            "get": {
                "tags": [
                    "AI Assistant"
                ],
                "summary": "Get AI-suggested questions",
                "description": "Get a list of suggested questions the user can ask the AI assistant based on their health profile and purchase history.",
                "operationId": "7ba6daf17535bd0bc3036a2f806e5dbf",
                "responses": {
                    "200": {
                        "description": "Suggested questions retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "suggestions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "What are the best foods for heart health?",
                                                        "Show me low-sugar snacks",
                                                        "Suggest a high-protein breakfast"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Suggestions retrieved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/ai/identify": {
            "post": {
                "tags": [
                    "AI Assistant"
                ],
                "summary": "Identify product from image",
                "description": "Upload a product image for AI-powered identification. Returns product details and nutritional information if found in the catalog.",
                "operationId": "d798fe986306805ed22a75010ff5cf16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product identified successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "identified_product": {
                                                    "type": "object"
                                                },
                                                "confidence": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.95
                                                },
                                                "nutrition_summary": {
                                                    "type": "string",
                                                    "example": "This product is high in protein and low in sugar."
                                                },
                                                "healthier_alternatives": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product identified successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid image or file too large"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "AI service error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/customer/ai/chat/history": {
            "get": {
                "tags": [
                    "AI Assistant"
                ],
                "summary": "Get chat history",
                "description": "Retrieve the authenticated user's AI assistant chat conversation history.",
                "operationId": "f48d1d2f11bd4eeb7b2944083c052a6d",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "parameter": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat history retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Chat history retrieved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Enter your API token ONLY (without 'Bearer ' prefix). Get this token from POST /api/auth/login response."
            }
        }
    },
...
}